+++ src/main/java/de/wbstraining/lotto/persistence/model/Users.java
@@ -43,7 +43,7 @@ public class Users implements Serializable {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Basic(optional = false)
     @Column(name = "user_id")
-    private Integer userId;
+    private Long userId;
     @Basic(optional = false)
     @NotNull
     @Size(min = 1, max = 10)
@@ -72,21 +72,21 @@ public class Users implements Serializable {
     public Users() {
     }
 
-    public Users(Integer userId) {
+    public Users(Long userId) {
         this.userId = userId;
     }
 
-    public Users(Integer userId, String username, String password) {
+    public Users(Long userId, String username, String password) {
         this.userId = userId;
         this.username = username;
         this.password = password;
     }
 
-    public Integer getUserId() {
+    public Long getUserId() {
         return userId;
     }
 
-    public void setUserId(Integer userId) {
+    public void setUserId(Long userId) {
         this.userId = userId;
     }
